refactor: move AnkiBroadcastReceiver to :common:android
The base class is a thin Context-locale wrapper around BroadcastReceiver
with no Anki/libanki coupling. Moving it down lets widgets, services
and other receivers extend it without depending on :AnkiDroid.
