+++ AnkiDroid/src/main/java/com/ichi2/widget/DayRolloverAlarm.kt
@@ -27,7 +27,7 @@ import androidx.core.app.PendingIntentCompat
 import androidx.core.content.getSystemService
 import anki.collection.opChanges
 import com.ichi2.anki.CollectionManager.withOpenColOrNull
-import com.ichi2.anki.android.AnkiBroadcastReceiver
+import com.ichi2.anki.common.android.AnkiBroadcastReceiver
 import com.ichi2.anki.common.coroutines.applicationScope
 import com.ichi2.anki.common.crashreporting.CrashReportService
 import com.ichi2.anki.common.exception.ManuallyReportedException
